Refactor Cytation: move all imaging logic to CytationMicroscopyBackend, use BioTekBackend directly
- Delete cytation.py (replaced by microscopy backend)
- CytationMicroscopyBackend now owns AravisCamera, camera_serial, imaging_config
- Device classes use BioTekBackend directly as driver (no more CytationAravisDriver)
- use_cam, camera_serial, imaging_config are device-level params passed to backend
- Extract _CytationBase for shared device logic (Cytation1/5)
- Add loading tray capability to both Cytation models
- Filter/objective loading happens in _on_setup
- CytationAravisDriver kept but no longer used by device classes
