+++ pylabrobot/agilent/__init__.py
@@ -1,12 +1,11 @@
 from .biotek import (
   EL406,
-  AravisImagingConfig,
   BioTekBackend,
+  BioTekLoadingTrayBackend,
   Cytation1,
   Cytation5,
-  CytationAravisDriver,
-  CytationBackend,
   CytationImagingConfig,
+  CytationMicroscopyBackend,
   EL406Driver,
   EL406PlateWasher96Backend,
   EL406ShakingBackend,
