+++ pylabrobot/agilent/biotek/__init__.py
@@ -1,8 +1,7 @@
 from .biotek import BioTekBackend
-from .cytation import CytationBackend, CytationImagingConfig
 from .cytation1 import Cytation1
 from .cytation5 import Cytation5
-from .cytation_aravis_driver import AravisImagingConfig, CytationAravisDriver
+from .cytation_microscopy_backend import CytationImagingConfig, CytationMicroscopyBackend
 from .loading_tray_backend import BioTekLoadingTrayBackend
 from .el406 import EL406, EL406Driver, EL406PlateWasher96Backend, EL406ShakingBackend
 from .synergy_h1 import SynergyH1, SynergyH1Backend
