+++ pylabrobot/legacy/plate_reading/agilent/biotek_cytation_backend.py
@@ -2,8 +2,8 @@
 
 from typing import Optional
 
-from pylabrobot.agilent.biotek import cytation
-from pylabrobot.agilent.biotek.cytation import CytationImagingConfig
+from pylabrobot.agilent.biotek import cytation_aravis_driver
+from pylabrobot.agilent.biotek.cytation_aravis_driver import CytationImagingConfig
 from pylabrobot.legacy.plate_reading.agilent.biotek_backend import BioTekPlateReaderBackend
 from pylabrobot.legacy.plate_reading.backend import ImagerBackend
 from pylabrobot.legacy.plate_reading.standard import (
@@ -18,9 +18,9 @@ from pylabrobot.resources import Plate
 
 
 class CytationBackend(BioTekPlateReaderBackend, ImagerBackend):
-  """Legacy. Use pylabrobot.agilent.CytationBackend instead."""
+  """Legacy. Use pylabrobot.agilent.CytationAravisDriver instead."""
 
-  _new: cytation.CytationBackend
+  _new: cytation_aravis_driver.CytationAravisDriver
 
   def __init__(
     self,
@@ -28,8 +28,7 @@ class CytationBackend(BioTekPlateReaderBackend, ImagerBackend):
     device_id: Optional[str] = None,
     imaging_config: Optional[CytationImagingConfig] = None,
   ) -> None:
-    # Override _new with the Cytation backend (which extends BioTek)
-    self._new = cytation.CytationBackend(
+    self._new = cytation_aravis_driver.CytationAravisDriver(
       timeout=timeout, device_id=device_id, imaging_config=imaging_config
     )
 
@@ -42,9 +41,9 @@ class CytationBackend(BioTekPlateReaderBackend, ImagerBackend):
     self._new.imaging_config = value
 
   async def setup(self, use_cam: bool = False) -> None:
-    from pylabrobot.agilent.biotek.cytation import CytationBackend as _NewCytationBackend
+    from pylabrobot.agilent.biotek.cytation_aravis_driver import CytationAravisDriver
 
-    await self._new.setup(backend_params=_NewCytationBackend.SetupParams(use_cam=use_cam))
+    await self._new.setup(backend_params=CytationAravisDriver.SetupParams(use_cam=use_cam))
 
   async def stop(self):
     await self._new.stop()
@@ -59,50 +58,50 @@ class CytationBackend(BioTekPlateReaderBackend, ImagerBackend):
 
   @property
   def objectives(self):
-    return self._new.objectives
+    return self._new.microscopy_backend.objectives
 
   @property
   def filters(self):
-    return self._new.filters
+    return self._new.microscopy_backend.filters
 
   async def close(self, plate=None, slow=False):
     await self._new.close(plate=plate, slow=slow)
 
   def start_acquisition(self):
-    self._new.start_acquisition()
+    self._new.microscopy_backend.start_acquisition()
 
   def stop_acquisition(self):
-    self._new.stop_acquisition()
+    self._new.microscopy_backend.stop_acquisition()
 
   async def led_on(self, intensity=10):
-    await self._new.led_on(intensity=intensity)
+    await self._new.microscopy_backend.led_on(intensity=intensity)
 
   async def led_off(self):
-    await self._new.led_off()
+    await self._new.microscopy_backend.led_off()
 
   async def set_focus(self, focal_position):
-    await self._new.set_focus(focal_position)
+    await self._new.microscopy_backend.set_focus(focal_position)
 
   async def set_position(self, x, y):
-    await self._new.set_position(x, y)
+    await self._new.microscopy_backend.set_position(x, y)
 
   async def set_auto_exposure(self, auto_exposure):
-    await self._new.set_auto_exposure(auto_exposure)
+    await self._new.microscopy_backend.set_auto_exposure(auto_exposure)
 
   async def set_exposure(self, exposure):
-    await self._new.set_exposure(exposure)
+    await self._new.microscopy_backend.set_exposure(exposure)
 
   async def select(self, row, column):
-    await self._new.select(row, column)
+    await self._new.microscopy_backend.select(row, column)
 
   async def set_gain(self, gain):
-    await self._new.set_gain(gain)
+    await self._new.microscopy_backend.set_gain(gain)
 
   async def set_objective(self, objective):
-    await self._new.set_objective(objective)
+    await self._new.microscopy_backend.set_objective(objective)
 
   async def set_imaging_mode(self, mode, led_intensity):
-    await self._new.set_imaging_mode(mode, led_intensity)
+    await self._new.microscopy_backend.set_imaging_mode(mode, led_intensity)
 
   async def capture(
     self,
@@ -116,15 +115,13 @@ class CytationBackend(BioTekPlateReaderBackend, ImagerBackend):
     plate: Plate,
     **kwargs,
   ) -> ImagingResult:
-    # Map legacy ImagingMode to new ImagingMode
     from pylabrobot.capabilities.microscopy.standard import ImagingMode as NewImagingMode
-
-    new_mode = NewImagingMode[mode.name]
     from pylabrobot.capabilities.microscopy.standard import Objective as NewObjective
 
+    new_mode = NewImagingMode[mode.name]
     new_objective = NewObjective[objective.name]
 
-    result = await self._new.capture(
+    result = await self._new.microscopy_backend.capture(
       row=row,
       column=column,
       mode=new_mode,
